Many scrolling speed improvements to SubmissionLink VC
